Prevent conflicting attributes from being paired
Many character qualities don't go well together. For example, a 
character who is an excellent swimmer is very unlikely to have a great 
fear of the water, and a person whose face is unreadable cannot 
simultaneously have a very dynamic face. A new array has been created, 
called 'exclusions', which is an array of multiple arrays, each of which 
contains a pair of arrays. For each pair, character attributes in the 
first array are exclusive of attributes in the second. To accompany 
'exclusions', the 'select' function now includes a while loop that 
checks whether two mutually exclusive attributes have been assigned to a 
character. If mutually exclusive attributes are found, one attribute is 
re-selected, and the loop checks again, completing only when all 
attributes are compatible.