Create form for re-sending verification email
If users do not verify their email addresses before the token expires, 
they must receive a new token. The form (EmailVerifyForm on 
'send_verify.html') takes only an email address, and upon verification, 
the provided address is checked to see whether verification is required, 
or if the email was never registered (in which case, the user is 
redirected to the sign-up page).