Prevent idle username invalidation
Allow usernames of unverified email addresses to be freed up if another 
user attempts to use the same username. For example, if a user 
accidentally types the wrong email address, that user can sign up a 
second time with the same username, and the old account will be deleted. 
Alternatively, if a user signs up with one username, then decides not to 
verify the email address, another user can claim the username. If, 
however, the email address has been verified, the username is 
unavailable.