Update redirects
Authenticated users are redirected from 'email_login' and 'verify_email' 
pages to 'home'. If a user tries to verify an email address while logged 
in as a different user, the current user is logged out and the address 
is verified, unless the token has expired. Email verification redirects 
to 'email_login'.