Enable password reset
On 'login.html', "Forgot your password?" appears under the password 
field. The text is an anchor that points to 'reset_password_request', 
which allows a user to enter the email address associated with the 
missing password (using the EmailVerifyForm). An email with a reset 
token is sent to the user's email address (if the email address is 
associated with a user in the database). The link containing the reset 
token points to 'reset_password', where the user enters and confirms a 
new password with the ResetPasswordForm. The new password is hashed and 
stored in the database, then the user can log in.