Allow users to edit account settings
Create new 'SettingsForm' in 'forms.py'. Add new 'temp_email' column to 
'User' model. On 'base.html', add link to new 'settings' view function. 
This new 'settings' view function renders the 'SettingsForm' on the new 
'settings.html' template. The form includes a FileField for profile 
pictures, but this is currently not rendered. Fields are rendered for 
name, username, and email address. The form validation includes checking 
email and username for uniqueness. Upon form submission, name and 
username are automatically added to the database, but a new email 
addresses must be verified through a link sent to the new email address, 
which is temporarily stored in the new 'temp_email' column. When the 
address is verified, the user's email is updated in the database. Upon 
verifying the new address, or upon logging in, 'temp_email' is set to 
'None'.