Allow users to add/update profile picture
Include 'profile_picture' field in 'settings.html', and ajust 
'settings.html' and 'main.css' to effectively display the current 
profile picture. The current profile picture is fetched from Dropbox 
using the new 'get_profile_picture' function, and a new one is sent with 
the new 'upload_profile_picture' function. As a new image is being 
uploaded, the database is being updated to note the name of the new 
image so that it may be fetched when needed.