Ensure that all profile pictures are 200x200
Formerly, images that were not perfectly square ended up too small since 
pillow would only resize the images, not crop them. Additionally, pillow 
only sizes images down, and will not blow up an image to the correct 
size. As a means of compensating, non-square images are now pasted onto 
a white 200x200 background after resizing. The center position for the 
image is calculated prior to pasting so that all images are as centered 
on the background as possible.