Modify 'Character' model
Formerly, all character data was stored as a dumped JSON in a column 
called 'data'. This was to hopefully make it easier to extend the 
Character model without having to add extra columns. However, other 
complications have made it necessary to use columns for every character 
attribute, and the model (and references to the model) has been updated.