Improve appearance of character saving
Formerly, after saving a character, the user would be redirected to the 
home page, and the new character (and flash message about the new 
character) would not show up until the next refresh or other navigation. 
Now, a JavaScript timeout delays the redirect for 200 milliseconds 
before sending the user to 'save_character', which then redirects the 
user to the appropriate character page. At the character page, the flash 
message shows up as expected.