Modify JavaScript redirect after saving character
Settle 'characterPromise' before redirecting, rather than waiting 200ms 
and hoping the promise was fulfilled in that timeframe. Upon successful 
POST, the promise is settled by logging the success message to the 
console and redirecting to the new character's page. On an unsuccessful 
POST, an error message is logged to the console.