Allow users to reset password from 'settings'
A user may follow an anchor on 'settings.html' to have a reset password 
URL sent to the user's email address. The URL points to 
'reset_password', just like the 'Forgot your password?' email. Upon 
submitting a new password, the user is logged out and sent to the 
'email_login' page, where the user may log in and then be redirected to 
'settings'.