Prevent users from setting 'email' to 'None'
If a user logs in after attempting to change email addresses in 
settings, 'temp_email' becomes 'None'. Formerly, a user could save a new 
email address, log in, then click the verification link and set 
'user.email' to 'None', which would probably prevent logging in forever. 
Now, if the verification URL is clicked, but 'temp_email' is 'None', the 
user is redirected to 'settings' with a flash message about the problem 
and nothing is saved.