Fix bug where profile picture would not load
Formerly, immediately after a user submitted a new profile picture, the 
image would not display properly until the page was reloaded. This was 
because a link to the old image was loaded on the POST request, then the 
old image was deleted, and no link to the new image was ever fetched 
until the next request. This was fixed by waiting to fetch the image 
until after form validation, so that the image link is fetched after any 
image changes are made. The related flash message apologizing for the 
delay has been removed.