Enable character editing
Add new 'Edit' button to 'character.html'. Button is embedded in anchor 
that points to the new 'edit' view, which renders 'new_character.html' 
with all table cells already filled with the character's data. The 
'Save' button will post the data to 'edit' instead of 'new_character' 
when accessed through the 'edit' view. The character is then updated in 
the database, and the user is redirected to the character's page. The 
'new_character' view and 'main.js' have been updated to ensure that the 
'new_character.html' only fills with random data if accessed through 
'new_character'.