Enable character deletion
Add new 'delete' view that accepts POST requests for character deletion. 
Users must be logged in, and a character may only be deleted by its 
corresponding user. The 'Delete' button on 'character.html' points to 
this view, but only posts after a confirm box asks the user to confirm 
the deletion.