Create Favorite characters
Allow users to create and add to a list of favorite characters. A new 
'Favorite' model has been created, with backrefs to both User and 
Character models. New 'add_favorite' view function queries a Character 
by id and creates a new Favorite instance using the character and the 
current user, but only if the character is not private. The 
'character.html' template has an anchor pointing to the 'add_favorite' 
view if the character is not private.