Create user profile page
Create 'users.profile' view function, which renders 'profile.html' 
template. The template displays up to 5 most recent characters and up to 
5 most recent favorites of the user, as well as the user's username and 
profile picture. Each character displayed is an anchor to the 
corresponding character page, and the end of the 'Recent Characters' is 
an anchor pointing to all the user's characters.