Allow users to "un-favorite" Characters
Create new 'remove_favorite' view function, which deletes the Favorite 
associated with the current user and a given Character, then redirects 
back to the Character's page. The 'character.html' template has been 
updated to display a button for removing a Favorite if the Character is 
already a Favorite. The button activates a JavaScript listener, which 
asks the user for confirmation before redirecting to 'remove_favorite'. 
The button only shows up if the Character is already a favorite, which 
is now determined in the 'character' view function and passed to the 
template.