Fix bug in determining user favorites
Formerly, an error occurred if an anonymous user tried to access a 
character page since the 'character' view function would try to 
determine whether the current user had added the character as a favorite 
while the query could not use an anonymous user. Now, the query only 
takes place if the current user is logged in.