Add print option
Create new 'print_character' view function, which renders 
'print_character.html' template, which displays the same information as 
'character.html' in a printer-friendly format. The template also 
displays a 'Print' button to open the print dialog (coded in 'print.js') 
and a 'Back' anchor that returns to the character page. The view 
function sends a boolean value (named 'print') to 'base.html' that turns 
off the device-width meta data so that the printer-friendly view appears 
the same on desktop and mobile. The 'character.html' template includes 
an anchor to the 'print_character' view.